# backend/docs/swagger.yaml
basePath: /api
definitions:
  gymtrack-backend_internal_domain_models.AuthorRole:
    enum:
    - trainer
    - athlete
    type: string
    x-enum-varnames:
    - AuthorRoleTrainer
    - AuthorRoleAthlete
  gymtrack-backend_internal_domain_models.Comment:
    properties:
      authorId:
        type: string
      authorRole:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.AuthorRole'
        enum:
        - trainer
        - athlete
      commentId:
        type: string
      content:
        maxLength: 2000
        minLength: 1
        type: string
      createdAt:
        type: string
      editedAt:
        type: string
      parentCommentId:
        type: string
      targetId:
        type: string
      targetType:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.TargetType'
        enum:
        - workout
        - meal
      type:
        description: Always "comment"
        type: string
    required:
    - authorId
    - authorRole
    - content
    - targetId
    - targetType
    type: object
...
  gymtrack-backend_internal_domain_models.TargetType:
    enum:
    - workout
    - meal
    type: string
    x-enum-varnames:
    - TargetTypeWorkout
    - TargetTypeMeal
...
  internal_api_handlers.CreateCommentRequest:
    description: Request body for creating a new comment on a workout or meal.
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
      parentCommentId:
        type: string
      targetId:
        type: string
      targetType:
        $ref: '#/definitions/gymtrack-backend_internal_domain_models.TargetType'
    required:
    - content
    - targetId
    - targetType
    type: object
...
  internal_api_handlers.UpdateCommentRequest:
    description: Request body for updating an existing comment's content.
    properties:
      content:
        maxLength: 2000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness tracking API for personal trainers and athletes
  title: GymTrack API
  version: "1.0"
paths:
...
  /comments:
    get:
      consumes:
      - application/json
      description: Retrieve all comments for a specific workout or meal. The user
        must have access to the target.
      parameters:
      - description: Type of target (workout or meal)
        in: query
        name: targetType
        required: true
        type: string
      - description: ID of the target (workout or meal)
        in: query
        name: targetId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comments retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing or invalid query parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Target not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve comments
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get comments for a target
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Create a new comment on a workout or meal. The user must have permission
        to comment on the target.
      parameters:
      - description: Create comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment created successfully
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.Comment'
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied or cannot comment on this target
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Target (workout or meal) not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create comment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - Comments
  /comments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing comment. Only the author can delete their comment.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only the comment author can delete it
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete comment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      consumes:
      - application/json
      description: Update the content of an existing comment. Only the author can
        edit their comment.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Update comment request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.Comment'
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only the comment author can edit it
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update comment
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - Comments
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"